+++ src/main/java/org/corpus_tools/annis/gui/components/codemirror/AqlCodeEditorState.java
@@ -17,6 +17,8 @@ import com.vaadin.shared.ui.JavaScriptComponentState;
 import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.Comparator;
+import java.util.HashSet;
+import java.util.Set;
 import java.util.TreeMap;
 import javax.xml.bind.annotation.XmlRootElement;
 import org.corpus_tools.annis.api.model.BadRequestErrorAQLSyntaxError;
@@ -92,6 +94,7 @@ public class AqlCodeEditorState extends JavaScriptComponentState {
     public final TreeMap<String, Integer> nodeMappings = new TreeMap<>(new StringComparator()); // NO_UCD
                                                                                                 // (use
                                                                                                 // default)
+    public final Set<Integer> optionalNodes = new HashSet<>(); // NO_UCD (use default)
 
     public String textareaClass; // NO_UCD (use default)